pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-jira-service-desk"
version = "0.1.0"
description = "MCP server for Jira Service Management — connect Claude Code, Claude Desktop, Cursor and other AI assistants to JSM"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "CzSadykov" },
]
keywords = [
    "mcp",
    "jira",
    "service-desk",
    "atlassian",
    "jsm",
    "claude",
    "claude-code",
    "model-context-protocol",
    "ai-tools",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
    "Typing :: Typed",
]
dependencies = [
    "atlassian-python-api>=3.41.0",
    "mcp>=1.0.0",
    "fastmcp>=2.0.0",
    "pydantic>=2.0.0,<3.0",
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/CzSadykov/mcp-jirasm"
Repository = "https://github.com/CzSadykov/mcp-jirasm"
Issues = "https://github.com/CzSadykov/mcp-jirasm/issues"

[project.scripts]
mcp-jira-service-desk = "mcp_jira_service_desk:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_jira_service_desk"]

[tool.hatch.build.targets.sdist]
include = [
    "src/",
    "tests/",
    "LICENSE",
    "README.md",
    "pyproject.toml",
    ".env.example",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "onpremise: tests specific to on-premise (Server/Data Center) deployment",
]
